+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Add Thomas chaotic attractor to 1-Op Chaos (https://sprott.physics.wisc.edu/chaos/symmetry.htm).
 - Make Mode knobs on 1-Op Chaos, 2-Op Chaos, Chaotic Attactors, and Clocked Noise snap to values and print the current mode.
 - Add "Low-frequency Oscillator" tag to Feedback Sine Chaos.
-- TODO: Make Scanner polyphonic.
+- Scanner is now polyphonic.
 
 ## 2.2.1
 - Add XOR feedback toggle to Rungler. Apply XOR feedback to locked and unlocked shift register.